requirements.txt
streamlit
pandas
plotly
folium
streamlit-folium
psycopg2-binary
pyotp
requests
geoip2
scapy
